Renamed ArmMotionMagic to MotionMagicArm.  Renamed Arm to BasicArm. I created the abstract Arm class, which BasicArm and MotionMagicArm extend.  I updated Robot and DriverInput to expect Arm so we can use either Arm type.  I cleaned up the Arm related constants in SystemSettings.
