+++ robot/src/main/java/us/ilite/robot/driverinput/DriverInput.java
@@ -16,7 +16,8 @@ import us.ilite.common.types.input.EInputScale;
 import us.ilite.common.types.input.ELogitech310;
 import us.ilite.robot.commands.Delay;
 import us.ilite.robot.modules.Arm;
-import us.ilite.robot.modules.ArmMotionMagic;
+import us.ilite.robot.modules.BasicArm;
+import us.ilite.robot.modules.MotionMagicArm;
 import us.ilite.robot.modules.Drive;
 import us.ilite.robot.modules.DriveMessage;
 import us.ilite.robot.modules.Module;
@@ -39,8 +40,7 @@ public class DriverInput extends Module {
 
     private Data mData;
 
-    //TODO: Change ArmMotionMagic back to Arm after testing
-    private ArmMotionMagic mArm;
+    private Arm mArm;
     private RangeScale armJoyStickToAngleScaler = new RangeScale(-1.0, 1.0, SystemSettings.kArmMinAngle, SystemSettings.kArmMaxAngle);
 
     public DriverInput(Drive pDrivetrain, Superstructure pSuperstructure, Data pData, boolean pSimulated) {
@@ -57,8 +57,7 @@ public class DriverInput extends Module {
         }
     }
 
-    //TODO: Change ArmMotionMagic back to Arm after testing
-    public DriverInput(Drive pDrivetrain, Superstructure pSuperstructure, Data pData, ArmMotionMagic pArm) {
+    public DriverInput(Drive pDrivetrain, Superstructure pSuperstructure, Data pData, Arm pArm) {
         this(pDrivetrain, pSuperstructure, pData, false);
         this.mArm = pArm;
     }
